Implement other less popular allocation functions
This is a large change. Most of the change is adding support for aligned
allocation. One change due to this also simplifies some other code: A
pseudo-allocation with size zero is created when the heap is
initialized. Since this will never be freed, it acts as a stopper when
freeing memory. It also simplifies the aligned allocation, which depends
on a prior allocation whose size can be increased to align the
subsequent allocation. A few other miscellaneous functions are also
added: malloc_usable_size, reallocarray, and reallocf.